[프로젝트 생성] 처음 생성 시 프로젝트 멤버가 있을때, 기능을 추가 했을 경우를 확장성 있게 변경

+++ src/main/java/com/lms/api/common/entity/project/ProjectFunctionEntity.java
@@ -24,7 +24,7 @@ public class ProjectFunctionEntity extends BaseEntity {
 
     String projectFunctionName;
 
-    Long projectFunctionSort;
+    int projectFunctionSort;
 
     @Enumerated(EnumType.STRING)
     ProjectFunctionType projectFunctionType;
